requirements.txt
Django>=2.1
python-dotenv
python-telegram-bot
psycopg2-binary
redis
uwsgi
pypeg2
faker
django-redis
django-cors-headers
django-debug-toolbar
django-filter>=2
celery[msgpack,redis]
djangorestframework
django-filter
django-rest-swagger
